Update and rename capitalism/index.md to draft/capitalism.md
my internal harmony says, now in 2019, after a few months of writing this for the first time and reviewing this now in December, that a huge change is on my way... and perhaps it goes through more layers of transparency, leaving such personal comments in the very last layer. which, even now, writing this here, is probably not ideal... but i'll keep improving!